Fix for #1890 (non-ascii characters in commit messages).
Request the encoding field from the commit. If no encoding is given,
default to UTF-8. Decode the body as well as author name and email
using said encoding.

This commit also reorders the fields in the log format a bit, and
moves author date to the second position, followed by the encoding.